Correct the cell locations, Change RPM to integer

+++ app/src/main/java/io/github/thymythos/diagnosticdataviewer/TableFragment.java
@@ -38,7 +38,7 @@ public class TableFragment extends Fragment implements LiveDataFragment {
     private GradientDrawable gdAmber;
     private GradientDrawable gdRed;
 
-    private float rpm = 0;
+    private int rpm = 0;
 
     public TableFragment() {
         // Required empty public constructor
